desktop-app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
crewai==0.36.1
crewai-tools==0.3.0
langchain-google-genai==1.0.6
langchain==0.2.0
python-dotenv==1.0.1
loguru==0.7.2
sqlalchemy==2.0.30
pysqlite3-binary==0.5.2
psycopg2-binary==2.9.9
chromadb==0.5.0
psutil==5.9.8
pydantic==2.8.0
pydantic-settings==2.2.1
jinja2==3.1.3 